Help Desk update
git-svn-id: file:///home/svn/DVA2006@3442 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/java/org/deltava/commands/help/IssueUpdateCommand.java
@@ -13,7 +13,7 @@ import org.deltava.security.command.HelpDeskAccessControl;
 import org.deltava.util.CollectionUtils;
 
 /**
- * A Web Site Command to update a Help Desk Issue. 
+ * A Web Site Command to update a Help Desk Issue.
  * @author Luke
  * @version 1.0
  * @since 1.0
@@ -21,55 +21,57 @@ import org.deltava.util.CollectionUtils;
 
 public class IssueUpdateCommand extends AbstractCommand {
 
-    /**
-     * Executes the command.
-     * @param ctx the Command context
-     * @throws CommandException if an unhandled error occurs
-     */
+	/**
+	 * Executes the command.
+	 * @param ctx the Command context
+	 * @throws CommandException if an unhandled error occurs
+	 */
 	public void execute(CommandContext ctx) throws CommandException {
 
 		try {
 			Connection con = ctx.getConnection();
-			
+
 			// Get the Issue
 			GetHelp dao = new GetHelp(con);
 			Issue i = dao.getIssue(ctx.getID());
 			if (i == null)
 				throw notFoundException("Invalid Issue - " + ctx.getID());
-			
+
 			// Check our access
 			HelpDeskAccessControl ac = new HelpDeskAccessControl(ctx, i);
 			ac.validate();
 			if (!ac.getCanUpdateContent())
 				throw securityException("Cannot update Comments/FAQ status");
-			
+
 			// Build a map of comments
 			Map<Long, IssueComment> comments = new HashMap<Long, IssueComment>();
-			for (Iterator<IssueComment> ici = i.getComments().iterator(); ici.hasNext(); ) {
+			for (Iterator<IssueComment> ici = i.getComments().iterator(); ici.hasNext();) {
 				IssueComment ic = ici.next();
 				comments.put(new Long(ic.getCreatedOn().getTime()), ic);
 			}
-			
+
 			// Update FAQ attribute from the request
 			i.setFAQ(Boolean.valueOf(ctx.getParameter("isFAQ")).booleanValue());
-			
+
 			// Start the transaction
 			ctx.startTX();
-			
+
 			// Get the DAO and save the issue
 			SetHelp wdao = new SetHelp(con);
 			wdao.write(i);
-			
+
 			// Determine what items to delete
 			Collection<String> deleteIDs = ctx.getParameters("deleteID");
-			for (Iterator<String> di = deleteIDs.iterator(); di.hasNext(); ) {
-				Long commentID = Long.valueOf(di.next());
-				if (comments.containsKey(commentID)) {
-					comments.remove(commentID);
-					wdao.deleteComment(i.getID(), new Date(commentID.longValue()));
+			if (!CollectionUtils.isEmpty(deleteIDs)) {
+				for (Iterator<String> di = deleteIDs.iterator(); di.hasNext();) {
+					Long commentID = Long.valueOf(di.next());
+					if (comments.containsKey(commentID)) {
+						comments.remove(commentID);
+						wdao.deleteComment(i.getID(), new Date(commentID.longValue()));
+					}
 				}
 			}
-			
+
 			// Determine what item to make the FAQ answer
 			Collection<String> faqIDs = ctx.getParameters("faqID");
 			if (CollectionUtils.isEmpty(faqIDs)) {
@@ -81,7 +83,7 @@ public class IssueUpdateCommand extends AbstractCommand {
 				if (comments.containsKey(commentID))
 					wdao.markFAQ(i.getID(), new Date(commentID.longValue()));
 			}
-			
+
 			// Commit the transaction
 			ctx.commitTX();
 		} catch (DAOException de) {
